mpidbg: replace MPIU_DBG_PRINTF with new format.
This is not exactly a one-to-one translation, since MPIU_DBG_PRINTF is
active only when MPICH_DBG_OUTPUT is true.  But MPIU_DBG_MSG_* are
active when USE_DBG_LOGGING is true.  However, there is no configure
option to enable MPICH_DBG_OUTPUT, so these debug messages would be
useless in the old format anyway.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>