threads: move error assertions to the MPIDU layer.
The MPIU layer for threads still checks for errors, but does not assert
for their success.  Instead, it now simply passes on the error to the
upper layer, which can assert on it or use it in a different way.

The error assertions are not consistent.  Some functions assert on
errors, while others simply pass them on to the upper layer.  We need to
decide one way or another and be consistent.  This patch does not fix
this issue.

Signed-off-by: Halim Amer <aamer@anl.gov>