mpidbg: isolate class usage.
Do not intermix usage of classes across layers, where possible.  For
example, utility functions are independent of devices or the MPI
layer.  They should not be using the same classes as the other layers.

There are still some cases where we are not completely strict in the
isolation of classes, but this should catch most cases.

Signed-off-by: Halim Amer <aamer@anl.gov>