valgrind: explicitly initialize bytes.
Valgrind seems to be confused when a structure has padding and we only
initialize the fields of the structure without touching the padding
bytes.  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=524488

This patch checks if we are running in valgrind mode and explicitly
initializes all bytes in such cases.  This adds extra overhead, but
since this is done only in valgrind mode, the common-case should not
be affected.

Signed-off-by: Yanfei Guo <yguo@anl.gov>