Disable 2-8 length type copy if alignment is required.
In RMA immed_copy routine, short data is copied as single type element
(e.g., 8 bytes is copied as uint64_t). This can result in bus error on
alignment strict platforms, because the data buffer might not be aligned
for such type access. This patch disables this copy for 2-8 bytes data
and uses memcpy instead on alignment-sensitive platforms. We note that 1
byte data is still copied as single uint8_t, because char does not
require alignment.