Separate input params packet header/data for packet handlers.
The original code did not separate packet header and data in receive
progress, but only gave packet handlers the pointer to entire receive
buffer. For accessing data behind of packet header, the handlers had to
calculate the start address of data and update buffer length also
for packet header part. This patch changes this processing to a clearer
way that passes the start address of packet header and that of data
through two input parameters, thus the packet handlers only need to
update the length of data part.

Signed-off-by: Pavan Balaji <balaji@anl.gov>