Avoid using macro for declaring global variables.
Originally some global variables are declared via macros as follows:

" #define VAR      \
     int var_1;    \
     int var_2;    \
     int var_3; "

and it is used in the code as follows:

" VAR; "

This causes additional semicolon at the end of global variable
declaration, which makes suncc compiler throws warnings.

In this patch, we avoid using marcro for declaring global variables,
but directly declare them in the code.

Signed-off-by: Pavan Balaji <balaji@anl.gov>