Restructure MPIR_Request to have kind-specific fields.
Depending on the kind of request, some fields are irrelevant.  Simply
including all fields into the request structure bloats up the size of
the request.  This patch moves the fields into kind-specific
structures out of which a union is created.  This seems to reduce the
size of the request by around 16 bytes.

For persistent requests, we separate out the "partner request" into
"real request" and "partner request".  The latter is only used within
the device and need not be exposed to the upper layers.

Signed-off-by: Wesley Bland <wesley.bland@intel.com>