Fix alignment issue in MPL_malloc.
When MPL memory tracing is enabled, additional header is added for every
allocated internal buffer. The start address of internal buffer is
calculated as buf=ptr+sizeof(header). Therefore, the header size needs
to follow alignment rules for all possible types in order to ensure the
start address of buffer is also aligned.

The previous code only used double[19] for such purpose, which was
incorrect. This patch fixed it by using the MPL memory alignment union.

Signed-off-by: Pavan Balaji <balaji@anl.gov>