ROMIO Onesided write aggregation enhancements to support file striping
The algorithm has been extended to support multiple calls for a single
one-sided collective write operation for a file system utilizing
striping.  The algorithm can now be called once for each segment of
data, a segment being defined as a contiguous region of the file which
has up to one occurrence of each stripe - the data for each stripe
being written out by a particular aggregator.  The parameters for
ADIOI_OneSidedWriteAggregation have changed slightly - pushing the
computation of the first and last file offsets to the caller to avoid
re-computation of this based on the offset list and an extra parm for
a striping parameters data structure.  The call from GPFS therefore
needed to change to support this new definition, but indicating that
no striping was enabled via the striping parameter.  Each call for a
striping file system effectively packs one striping unit of data into
the collective buffer on each agg, with additional parameters which
govern flushing the collective buffer to the file.  In practice the
collective write for a file system such as lustre on a dataset
composed of multiple segments would call the algorithm several times
without a flush parameter to fill the collective buffers with multiple
stripes of data, before calling it again to flush the collective buffer
to the file system.  In this fashion the synchronization can be
minimized as that only needs to occur around the actual read from or
write to the file system.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>