Generalize ROMIO collective aggregation tuning
Use the TUNEGATHER environment variable to select different collective
tuning approaches.  The current implementation for GPFS allows the user
to use one MPI_Allreduce instead of 2 (or 3 in the case of onesided
aggregation) MPI_Allgathers to communication all starting and ending
file offsets (and onesided count sizes) if the GPFSMPIO_TUNEGATHER
environment variable is set.  Generalize to other file systems as
ROMIO_TUNEGATHER.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>