Make MPIR, MPIU, MPID, MPII, MPICH namespace consistent.
MPIR should only be used for functionality that is exposed by the
MPI-layer downward to the device.  Other functionality owned by the
MPI layer that is used internally within that layer should be called
MPII.  MPID functionality is device-specific functionality that is
exposed to the MPI layer.  MPIU namespace is being removed for now.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>