Fixes intercom_merge temporal context_id usage
This patch fixes a problem in the intecom_merge routine. Intercom_merge
needs a temporaly communicator to find the context_id for the merged
communicator. However, this communicator also needs a context_id,
which is unique on every process and the same on all processes.

The problem with the previous routine is that the context_id  of an
inter-communicator may be used by another communicator – which can be
an intra-communicator with a sub-internode communicator. In this case,
the context id is not longer unique. A subcommunicator with ID 3 normally
is not allowed - so it is  allways unique. But we have to make sure that
for this communicator, no sub-communicators are created, so we avoid
the creation of sub communicators, if the sub communicator flag is set.

Signed-off-by: Yanfei Guo <yguo@anl.gov>