Inline object allocation routines
Moves the object allocation routines into header files as static inlines.
In performance profiling of these routines, we noticed the instruction cost
of the function call is on the order of the cost of the object allocation
(~18 instructions for obj_alloc with the gnu 4.9 compilers).  A compiler
may still out of line these instructions, but IPO is not required to inline
them.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>