Disable vc state change in tcp if vc started closing.
Tcp netmod updated vc to ACTIVE state when connection is established.
It can happen in the first message. However, if the first message is vc
closing packet, that vc can never be closed. Because, during vc closing,
process needs do handshake with remote peer and change state:
	LOCAL_CLOSE/REMOTE_CLOSE -> CLOSE_ACKED -> CLOSED -> INACTIVE
The next state change depends on the current state. If such state is
changed LOCAL_CLOSE->ACTIVE in netmod, the closing process can never
finish. This patch fixed it. Tcp netmod only updates state when the vc
is INACTIVE.

Signed-off-by: Yanfei Guo <yguo@anl.gov>