Add debug support to persistent send
When the MPIR_Request object changed to a union, starting a persistent
send became impossible if debugger support and error checking is turned
on. This is because the union for the `dbg_next` field in the send
shares an address with the `real_request` field in the persistent part.
This fails the check later when calling `MPI_Start` because
`real_request` is not null so MPICH thinks the request has already been
started.

Change the union to include a `dbg_next` field in both regular and
persistent sends and check to make sure that only the correct one is
getting filled in.

Fixes csr/mpich-opa#336

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>