ROMIO Onesided write aggregation utilization for Lustre
The collective write aggregation call for Lustre
(ADIOI_LUSTRE_WriteStridedColl) has been enhanced to utilize
ADIOI_OneSidedWriteAggregation if the one-sided environment
variables are set.  The general algorithm is to divide the file up into
segements, a segment being defined as a contiguous region of the file
which has up to one occurrence of each stripe - the data for each
stripe being written out by a particular aggregator.  Iteratively call
ADIOI_OneSidedWriteAggregation for each segment to aggregate
the data to the collective buffers, but only do the actual write
once the appropriate number of stripes have been packed (based
on the size of the collective buffer) or the aggregation for all the
data is complete, minimizing synchronization.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>