Symbol visibility
This patch implements shared library best practice with regard to symbol
visibility. MPICH is compiled to hide all function and variable symbols,
except those decorated with MPICH_API_PUBLIC. Compiler support is tested
for, and visibility disabled if not available. A configure flag,
--enable-visibility, is added to control this option. Note that some
symbols must be exposed from libmpi.so that are not obviously user
visible, but are needed by the language binding libraries.