CH4/OFI: Add support for scalable address vector
Attempt to open a shared, readonly, address vector that has been
previously constructed by an external component, such as a proccess
manager or control system. If the open is successful the fi_addr_t
entries are copied out of the mapped table, otherwise the PMI KVS
exchange algorithm is use to initialize a new address vector.