CH4/OFI: Add configure time capability sets
Add a static list of capability sets based on the input provided at
configure time. If the user provides a set name to the `--with-device`
flag, it will be included here. Based on the set chosen, various OFI
features will be enabled or disabled at compile time to reduce branches
at runtime.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>