CH4: Add UPID and LUPID for intercomm map exchange
1. Unique PID (UPID) is the process network address. It is globally
unique and can be used for address exchange between different process
groups.

2. Local Unique PID (LUPID) is a unique identifier for processes within
the space of one process. It is essentially a packed format of (avtid,
lpid). A LUPID is currently 32-bit integer.

3. The map exchange is now handled at CH4 layer (with some help from
netmod). The intercomm_create can detect if the two groups belongs to
the same MPI_COMM_WORLD. If only processes in MPI_COMM_WORLD is
involved, the map exchange only need LUPIDs. If not, a UPIDs exchange is
reuquired, and each process need to covert UPIDs to LUPIDs locally
(handled by netmod).

4. Support for arbitrary length of UPID. CH4 allows a generalized UPID
format that allows variable UPID length. Note that the UPID handling is
implemented by netmod. One netmod can internally optimize if it knows
the UPIDs have the fixed length.

This patch introduced new interface. All netmods and shm modules are
updated accordingly.