Bug-fix: add MPID_Thread_yield.
We had initially missed adding MPID_Thread_yield and instead defined
MPID_Thread_same twice.  It's surprising that the compiler did not
catch the redefining.  It's also weird that we do not use
MPID_Thread_yield any where, but it's still exposed, though that part
is not fixed in this commit.

No reviewer.