Bug-fix: Correctly reference count datatypes
When an isend/issend/irsend returns without completion for derived
datatypes, we need to increment the datatype ref count
unconditionally.  We were checking to see if OnDataAvail was set
before doing that, which is incorrect.

The ref-count increase should only be for derived datatypes, and not
for builtin datatypes (we treat non-builtin predefined datatypes the
same way as derived datatypes).  The original check for issend was
incorrect in this respect.  This patch fixes that part too.

Fixes #2353.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>