CH4/OFI: Avoid messing up compile/link flags
The link flags for libfabric were getting messed up if the user passed
in embedded as the value because they were being added to LDFLAGS too
early. Avoid this by only doing it when we determine that we are not
embedding.

Also avoid nuking too much of the user inputed flags so we can keep
things like visibility preserved.