romio: initialize some global variables
Clang (clang++) on Mac OS complains that those variables are Undefined
symbols at linking time. It seems that uninitialized global variables
are stored in a common section and are not visible to the user program
at linking time. When initialized, Clang put them in the initialized
data section and become visible.