Remove the 32B alignment for MPID_Request
The MPID_Request structure is aligned to 32B and allocated through the
request memory object pool. The alignment was introduced for
potentially reducing false sharing in thread_multiple mode. Indirect
blocks in the pool, however, are dynamically allocated without
using aligned memory allocation. This does not guarantee 32B
alignment and causes segfaults when touching requests in indirect
blocks. To this point, there is no evidence of significant performance
improvement with this alignment. Thus this patch drops this alignment.

Fixes #2350

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>