Cleanup request creation/freeing functions.
Cleaning up of AM-related of partner-request-related resources can be
done as a free hook.  There is no need to do a separate wrapper
function.

For request creation, we need a wrapper in order to know when to
allocate the AM structures, but we can merge them into a single
function.  We had two versions of the AM request creation function,
one that sets the ref-count to 1 and another that sets the ref-count
to 2.  We merged these two functions and now take the ref-count as an
argument.