timers: fix and make optional the PPC64 cycle counter
If the user selects the PPC64 cycle counter, configure does not test
for it properly and fail gracefully. Furthermore, testing for this
counter is always performed by default (when the user did not ask
explicitly for a timer type). One certain platforms (e.g., Intel
compilers on OSX), this mandatory check breaks subsequent configure
tests unintentionally. This patch makes this cycle counter a
user-supplied option only (consistent with the other cycle counters)
and fails gracefully when configure cannot compile the corresponding
assembly.  It also avoids the default check.