Use rec_count 2 + request_complete for completed req in RMA.
If OPs are immediately issued, we need create a "completed request" for
request-based RMA OPs, thus user can wait on that request.

The old code called set_completed to avoid adding ref_count, but the
set_completed should not be called. In this patch, we set the request's
ref_count to 2, and then we can request_complete on it.

In future optimization, we should have a predefined request handler for
such special request.

Signed-off-by: Pavan Balaji <balaji@anl.gov>