CH4/OFI: Allow runtime configuration of capability sets
Allow the user to configure a capability set at runtime by selecting the
runtime capability set (or no capaiblity set at all) and filling in the
CVAR to pick a particular provider. If the user does this, the initial
capability set values will be pulled from the appropriate provider
(specified by using the MPIR_CVAR_OFI_USE_PROVIDER CVAR) and
can be customized by other CVARs.

Getting this to work requires a bit of a hack because by their nature,
capability sets must be visible at compile time, while runtime
configuration requires something available outside of just compiler
macros.  This means that some of the definitions had to be replicated
in both ofi_capability_sets.h and globals.c. However, the values themselves
are not replicated and changing them only in ofi_capability_sets.h is
safe. The important bit of state to preserve is the ordering of
capability sets in both locations.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>