CH4/OFI: Move tagged checks to runtime
Moves the checks for whether to use the tagged interface to the runtime
instead of compile time. While this seems like a bad idea, in reality,
it will be a branch that's optimized out because it will be a constant
when a capability set is selected. Only when using runtime selection
will there be an actual branch inserted.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>