CH4/OFI: Reduce immediate data field to 32 bits
The OFI spec only requires providers to supply 4 bytes for
`cq_data_size`. We allocate 64 bits for the dest argument despite the
fact that we never use more than 32 bits.

Change this to use a uint32_t for the dest argument so we don't overrun
the provider's CQ data field.

Fixes pmodels/mpich#2478

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>