CH4/UCX: Set source and tag status values in truncation case
The MPI standard encourages setting the source and tag information in
status for truncated messages. UCX itself does not guarantee the
content of its completion info in the case of truncation errors, but
passing it along can do no harm. In the event they are accurate, they
can be of use. See tests in test/mpi/errors/pt2pt.

No reviewer.