MPID_Abort fixes.
The previous fixes to MPID_Abort were causing it to fail to build on a
mac.  This patch allows MPID_Abort to be defined in a C file, with a
prototype in the overall header to allow all files to access it
correctly.

This fix is not yet complete.  The build still fails inexplicably if
the MPID_Abort is moved to a different file.  But at least it gets the
build going for now.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>