CH4/OFI: Fix match_bits with immediate data disabled
The macros to check if immediate data was enabled or not were not
checking properly and therefore always assuming that it was enabled.
This meant that the layout of the `match_bits` was sometimes incorrect
and caused bits to be misinterpreted.

As a side effect, this meant that I could fix the runtime layout so it
wouldn't have to default to the "safe" no-data version. Now it can do
runtime detection and pick the correct one for whatever the user picks.

Fixes csr/mpich-ofi#481

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>