CH4/OFI: Move direct provider flag from netmod args
Remove the configure flag for building with a direct provider from the
netmod args as the netmod args were otherwise unused. Create a new flag
that combines both previous flags to set up a direct provider.

One of these flags was specific to an embedded build of OFI. Now we'll
just use the new flag for both to simplify things.

Fixes pmodels/mpich#2518