CH4/OFI: Provide hints to users about ranks/tags
If the user asks for debugging information about the capability sets,
provide the maximum tag and number of ranks that can be supported.

There was also a discussion of providing the maximum number of
communicators, but this was left out for a few reasons:

The number of communicators is not the same as the number of context IDs
and therefore could confuse users.  The number is not as simple to
calculate due to the maximum number of bits not being available
currently (could be changed).  This number doesn't change based on the
capability sets. Right now, this number is always 2^8.  This patch will
also abort the job if the user asks for more ranks than MPICH can
support. This happens regardless of whether debugging is turned on or
not.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>