CH4/OFI: Specify send and receive capabilities
When the caps do not explicitly say if it will be for send or receive,
both are set. This causes problems because receive capabilities can be
applied to tx contexts and vice versa. By explicitly setting the caps,
we ensure that only the appropriate caps are set on corresponding
contexts.

Fixes csr/mpich-ofi#482