endpoints: first prototype with MPI_Send
This adds 1) a new CS granularity called "per-endpoint", 2) and tests it with
MPI_Send. It does not break other global and per-object granularities, reuses
the global CS to gradually move the thread safety to the device, and takes
advantage of some per-object locks, like the progress mutex, as a temporary
solution.