CH4/OFI: Convert macros to static inlines for ep lookup
For functions where the layout of the data structure can change
depending on the capability set configuration, we need to have more
ability to check macros that we can get within a macro definition.

Convert these macros to static inline functions so we can check all the
possiblities and remove compile errors when not using runtime sets.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>