pm/hydra: cleanly separate the rmk from the launcher.
The RMK is mostly independent of the launcher.  The launcher only
needs to know where the node list came from, but does not need any
real access to the RMK.  This patch creates a new copy of the RMK, but
leaves the RMK within the launcher as-is.  Separating the RMK from the
launcher is more error prone, so we'll do that as a separate commit.