pm/hydra: demux splicing code.
Code to forward data from one socket to another.  This is similar to
the splice command in linux, but is done in user space.  Also, this
code allows us to do an unstructured forwarding from multiple input
sockets to a single output socket.  It is "unstructured" in the sense
that the data from different sockets might be arbitrarily interleaved.