CH4/OFI: Fix huge message underflow
If the sender underflows the huge message protocol by not sending enough
data to trigger the protocol on the sender side, we need to know to bail
out on the receiving end.

Do some checks to avoid the problem by switching protocols on the
receiving side if the message size is less than
`MPIDI_Globals.max_send`.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>