Object: fix refcount verification logic
The current way of verifying refcount is:
  1: atomic_incr/decr(refcount);
  2: assert(atomic_load(refcount) >= 0); (in HANDLE_CHECK_REFCOUNT)
So there may be a race between line 1 and 2. (Suppose other thread
comes in and modifies refcount there.)

This patch provides a fix to this, by using fetch_and_atomic OPA
functions to retrieve refcount, and passing that value to the
verification macro so that it doesn't read the value again.

However, for non-debug builds, we still use non-fetch atomics,
because they may be slightly faster than fetch versions, and we don't
care about exact value of the refcount (other than whether it hit zero.)

Signed-off-by: Halim Amer <aamer@anl.gov>
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>