test/spawn: Avoid potential deadlock in join test
Completion of MPI_Send only guarantees that the user buffer is safe to
reuse. Before entering into the accept call in this test, we need to
ensure that the data sent with MPI has been received at the
target. We use MPI_Barrier in this case to make sure the MPI_Recv calls
have also completed at the target, and avoid potential deadlock.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>