ROMIO: Fix for ADIOI_OneSidedReadAggregation
There was a corner-case issue with the ADIOI_OneSidedReadAggregation
function where if the file region being read by a rank spanned multiple
file domains including the last one it could result in a hang or crash.
The issue was fixed as well as trace messages updated to better exhibit
the issue.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>