CH4/OFI: MPIDI_OFI_do_iprobe needs to set msg.data to source
In the MPIDI_OFI_do_iprobe function the data field in the msg passed to fi_trecvmsg is
currently just set to 0.  Instead this needs to be set to the source rank because for
MPIDI_OFI_ENABLE_DATA mode the MPI_Status.MPI_SOURCE field is set to the fi_cq_tagged_entry
data field via MPIDI_OFI_cqe_get_source -- this source rank needs to make it into the data
field in the completion queue entry.  The only reason that the pt2pt/mprobe test works is
because it is a 2-rank test and the source rank just so happens to be 0.