Fix load imbalance in implementation of Rabenseifner’s algorithm
In the current implementation of Rabenseifner’s algorithm for Reduce and Allreduce
there is a load imbalance across the processes. The reason is non uniform  distribution
of count elements across pof2 processes on the reduce-scatter and {all}gather steps.
The last process pof2-1 is loaded more than others:

For example,
    commsize=13, count=13; pof2=8, cnts[0..6]=1, cnts[7]=6.
    commsize=1600, count=100000; pof2=1024, cnts[0..1022]=97, cnts[1023]=769.

This implementation does not use cnts[i] and disps[i] arrays, so memory consumption per process
is reduced from O(2^\floor{\log(p)}) to O(\log(p)). In addition, the need to calculate cnts[i]
on each step of the reduce-scatter is eliminated.

Signed-off-by: Mikhail Kurnosov <mkurnosov@gmail.com>