ch4/ofi: Move OFI version into capability sets
In general, when optimizing, we want to use the latest and greatest
version of the libfabric library because we need things like
FI_CONTEXT2 or FI_RESTRICTED_COMP. On the other hand, when using runtime
capabilities, we'd like to maintain as much backward compatibility as
possible.

To do this, the compile time capability sets will specifiy a specific
libfabric version and the runtime capability set will specify libfabric
1.0 for now. This can be upgraded later as we drop support for older
versions of libfabric.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>