CH4/OFI: Define FI_MAJOR/MINOR_VERSION if unused
dca4794ad87da1f94cab5b02e4eace7efbb1cab5 uses `FI_MAJOR_VERSION` and
`FI_MINOR_VERSION` as the seed value for a couple of CVARs which means
that they are required to be defined, even if we're not using libfabric.

If the user doesn't specify that they want to use libfabric, specify
some throwaway values to make sure we can still compile.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>