CH4: Fixed win_alloc_shared when symm heap is unsupported.
In win_allocate_shared, we generate random address on rank 0 and force
all local processes to mmap with this address if symmetric heap is
supported. This requires mmap with MAP_FIXED. However, when symmetric
heap is not supported, we should not set such flag but should allow the
system to pick up one. Previous code did not cover the latter case. This
patch fixes it following the rules:

On rank 0:
1. If symmetric heap is enabled, try generate random address. If valid
   address is returned, then mmap with MAP_FIXED on rank 0.
2. If symmetric heap is disabled, or invalid address is returned, then
   mmap without MAP_FIXED on rank 0.

On other local ranks:
1. If symmetric heap is enabled, and the received address is also locally
   valid, then mmap with MAP_FIXED.
2. If symmetric heap is disabled, then mmap with the address but without
   MAP_FIXED.
3. If symmetric heap is enabled but address is locally invalid, then
   assert as original code does (this should be fixed in separate patch).

Fixes pmodels/mpich#2646

Signed-off-by: Hajime Fujita <hajime.fujita@intel.com>