autogen: avoid running find on the whole tree
In order to detect whether find and xargs are supported on the target
platform, autogen.sh runs the find command on the whole MPICH
directory ("."). This is overfill as we only want to detect that those
features exist. On some platforms (e.g., with NFS), this step can take
very a long time. This patch solves this issue by reducing the scope
of find to the ./maint directory.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>