Fix and clean up MPIR_Grequest_progress_poke
MPIR_Grequest_progress_poke had unused variables, might block in wait_fn (which
is incorrect), and assumed erroneously request_req[i+1] is a generalized request
because request_ptr[i] is. This patch cleans up this routine and makes it
nonblocking, which is required by some callers (e.g., Waitsome). The class-wise
completion optimization was given up in favor of correctness. A correct
implementation of the class-wise completion is left as future work.