Avoid erroneously blocking within Waitsome and Waitany
Waitsome and Waitany might wait for the completion or either native requests or
generalized requests. Since the progress engines for these two categories of
requests are separate, blocking in either one of them might hang because the
other request category might be the one that has some completed request. In
particular, the current implementation might block in the progress engine of
native requests while the pending requests are generalized ones. This patch
makes Waitsome and Waitany poll for progress on both type of requests in a
nonblocking manner.