fixup: Store individual ranks in the tree data structure
Fixup "coll: New collectives framework at mpi layer"

Storing rank ranges instead of individual ranks reduces transient memory
requirements but it complicates the implementation of some algorithms.
We are now storing individual ranks in the child list. This is fine
because the space requirement is logarithmic to the number of ranks,
which is small even for large number of ranks and also this is transient
space requirement.