coll: Add second type of knomial tree
The two tree types are called knomial_1 and knomial_2. The
difference between the two trees is the different assignment of
rank to vertices in the tree. knomial_2 tree gives the
same binomial tree as used by the MPIC_DEFAULT algos.

Also add bcast and reduce algorithms that use this new tree.