mpir: Add parens to ensure cast
Coverity doesn't seem to think that this code is good enough to ensure
that we're not bitshifting a 32-bit value (potentially count_) 32 bits:

    ((MPIR_Ucount) count_ >> (8 * SIZEOF_INT) << 1)

By adding an extra set of parens, this should convince coverity that
this operation is safe.

Alternatively, this patch could go and change all of the places where we
pass a 32 bit value into `MPIR_STATUS_SET_COUNT` (which is actually a
pretty small number), but that seems like the "wrong" solution to me.

(Possibly) Fixes #160910, #160911, #160912
Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>