coll: Maintain start_vtcs list in graph schedule
These are the vertices that have no incoming dependencies. The list is
determined when the schedule is committed. When a schedule begins
execution just issue all the vertices in this list instead of having to
traverse over all the vertices and issuing the ones with no incoming
edges. This optimization is beneficial for persistent collectives as
start_vtcs list needs to be determined only once.