ch3/ofi: Do not refcount temporary request
Probe tests occasionally segfault on an invalid refcount in this
configuration. This can happen because we would unconditionally add a
reference to the request here, despite the fact it may have been
allocated on the stack and thus be not fully initialized. Add a check
to make sure this is a valid request before referencing.

Signed-off-by: Halim Amer <aamer@anl.gov>