refactor MPID_Node_id_t objects to int type
There is no reason for the node_id to be different from that of an
integer.  It is not sent out on the wire, so the size of the datatype
does not matter.  And it cannot be larger than INT_MAX since the MPI
standard defines the number of ranks (and hence the nodes) to be
within INT_MAX.

Signed-off-by: Yanfei Guo <yguo@anl.gov>