ch4/progress: replace cs_exit + cs_enter by cs_yield
CH4 calls explicitly cs_exit and then immediately cs_enter inside progress_test
to avoid threads blocking others. Since intention was to yield the critical
section, thus this patch replaces those two calls by cs_yield. This is also a
prerequisite before using low priority locking inside progress routines, which
relies on cs_yield to detect the inability of a thread to progress.

Note that this does not fix the inherent progress issue in CH4, which maps
progress_wait to progress_test. Since progress_test is supposed to be
nonblocking, yielding inside it is unnecessary. The true fix, however, is more
elaborate and requires changing progress_wait instances with blocking on a
condition, in which case yield would be necessary.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>