comm: fixup case where user passes MPI_UNDEFINED
Earlier we were relying on the fact that all processes eventually call
MPIR_Comm_split_impl over the entire communicator, even if the device
overrides the function for some processes and hands it off to the
upper layer for other processes.

For example, with ch3:nemesis, in the shared memory case, all
processes with MPI_UNDEFINED were calling the upper-level
MPIR_Comm_split_impl, while the processes that gave
MPI_COMM_TYPE_SHARED would use the nemesis code.  The nemesis code
coincidentally happened to call MPIR_Comm_split_impl over the user
communicator as well, so this bug never surfaced.  But it is easy for
us to misstep and create a deadlock in the future.

In this patch, we simply split out "UNDEFINED" processes from the
parent communicator and handle the rest of the processes together.  If
the device overrides split_type, it would be responsible for dealing
with MPI_UNDEFINED.