pm/hydra: improve hydra nodelist parsing functionalities using POSIX regex
Fixes a problem reported in pmodels/mpich#2786. More specifically, slurm's
nodelist parsing in hydra was not able to recognize strings formatted as:
host-[001-012]. This commit adds support for such string format using POSIX
compliant regular expressions. It also provides hydra with all the auxiliary
configure checks, allowing it to fallback to the original implementation if
regcomp is not supported.

Signed-off-by: Pavan Balaji <balaji@anl.gov>