confdb: add a new check for F08 support
The new check is for the case where a async subarray is passed to a
assumed-type assumed-rank async dummy argument of a interface in a
module. ifort 18 is known to give compile error on it.

This is a patch to confdb after the switching to submodule. Here is to
bring it back.

Signed-off-by: Yanfei Guo <yguo@anl.gov>