pmi: bug fix in detecting gethostbyname.
Our general logic in searching for gethostbyname has typically been to
do AC_CHECK_FUNCS for gethostbyname, and if that is not found to do
AC_SEARCH_LIBS to find it in libnsl (mostly for solaris).  In some
cases, we were doing the opposite, i.e., AC_SEARCH_LIBS first and then
AC_CHECK_FUNCS.  This in itself is fine, but we were running into a
corner case where configure was picking up a previously cached value
of whether gethostbyname was found or not (before libnsl was added to
LIBS) and reporting that as a failure in finding gethostbyname.

This is not an hwloc issue, per se, but it showed up after we
integrated hwloc into mpich.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>