coll: update bcast algorithms for SMP-bcast.
Our current algorithm-selection for SMP-aware bcast is pretty ad hoc.
Some of the decisions are made based on the number of processes in the
parent communicator, which is no indication of the number of processes
in the subcommunicators.

This patch allows us to go through the regular collective algorithm
selection logic for the subcommunicators as well.

The following changes were made for the algorithms when the device
does not override bcast:

 - For medium sizes messages and if the total number of processes is
   large and a power of two:

     - when the number of processes on each node is large and is a
       power of two: change binomial to scatter_doubling_allgather to
       node_comm subcommunicator.

     - when the number of processes on each node is large and is not a
       power of two: change binomial to scatter_ring_allgather to
       node_comm subcommunicator.

     - when the number of nodes is small: change
       scatter_doubling_allgather or scatter_ring_allgather to
       binomial.

Signed-off-by: Pavan Balaji <balaji@anl.gov>