ch4/ofi: Rename cq static and dynamic lists
There are two types of lists for OFI completion queue events. There is a
statically allocated list for a small number of completion queue events
and a dynamic list when the static list is overrun.

The naming of these lists is confusing so here we updated the names of
the variables associated with those lists to make it clear when the
static list is being referenced and when the dynamic list is being
referenced.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>