coll: Store power of 2 in MPIR_Comm
Collective communication operations use nearest power
of 2 calculation for algorithm selection and also for
algorithm execution. Some code paths calculate it twice during
the course of a call to collective operation. Calculate it
just once and store it in the communicator. Retrieve it from the
communicator whenever needed instead of calculating
it on the spot everytime. This will save some performance
cost. pof2 is stored just after the local_size
parameter inside the communicator. local_size is anyways
accessed during collective execution. Therefore, chances
of a cache miss on pof2 are small.

Signed-off-by: Pavan Balaji <balaji@anl.gov>