ch4/generic: Move params inside each case in startall fn
As we add more cases (for example, for collectives) to
the switch statement, many of these parameteres
(like tag, datatype, rank, etc.) would no longer
be common across the cases in the switch statement. Hence, it
does not deem fit to calculate them for every case. Get
these variables directly inside each case. This also
makes it consistent with the way ofi_startall is
implemented.

Signed-off-by: Pavan Balaji <balaji@anl.gov>